+++ CHANGELOG.md
@@ -51,6 +51,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `--fee-token` flag - `strk` will be used by default
 
+### Cast
+
+#### Added
+
+- Voyager API support for sncast `verify` subcommand 
+
 ## [0.34.0] - 2024-11-26
 
 ### Forge